weather_alert/requirements.txt
aget==0.2.0
altgraph==0.17.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.5
click==8.1.8
cryptography==43.0.3
Cython==3.2.4
fastapi==0.128.8
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.1.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
packaging==26.0
passlib==1.7.4
pillow==11.3.0
pluggy==1.6.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.3
pytest==8.4.2
python-dotenv==1.2.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
ruff==0.15.5
setuptools==82.0.0
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.39.0
